requirements.txt
python-dotenv == 1.1.0
langchain == 0.3.25
langchain-community == 0.3.23
langchain-groq == 0.3.2
streamlit == 1.45.0
